temp_reference_code/backend/requirements.txt
# Web框架
fastapi==0.104.1
uvicorn[standard]==0.24.0
gunicorn==21.2.0
waitress==3.0.0  # Windows兼容的WSGI服务器

# 数据库相关
SQLAlchemy==2.0.23
PyMySQL==1.1.0
cryptography==41.0.7
alembic==1.12.1

# 认证和安全
passlib[bcrypt]==1.7.4
python-jose[cryptography]==3.3.0
python-multipart==0.0.6
pycryptodome==3.19.0

# 数据验证
pydantic==2.5.0
pydantic-settings==2.1.0

# HTTP客户端
httpx==0.25.2
requests==2.31.0

# 日期时间处理
python-dateutil==2.8.2

# 配置管理
python-dotenv==1.0.0

# 日志
loguru==0.7.2

# 缓存
redis==5.0.1

# 文件处理
Pillow==10.1.0
python-magic==0.4.27

# 工具库
click==8.1.7
rich==13.7.0

# 开发工具
pytest==7.4.3
pytest-asyncio==0.21.1
black==23.11.0
flake8==6.1.0
mypy==1.7.1

# API文档
fastapi-users==12.1.2

# 跨域支持
fastapi-cors==0.0.6

# 环境变量
pydantic-settings==2.1.0

# JSON处理
orjson==3.9.10

# 异步任务
celery==5.3.4

# 监控
prometheus-client==0.19.0
psutil==5.9.6

# 对象存储
minio==7.2.0